.zshrc: Install fzf package
fzf is a handy command line fuzzy finder, much like HELM embedded in a
shell. To whom it may concern: https://github.com/junegunn/fzf

Surely, this lacks the vital part here: automation of installation
itself, but this can be achieved as simple as `pacman -S fzf` on
Arch Linux distros and by cloning the repo and running an install script
on any possible *nix system (I suppose, so to say).